Introduces first separation of NetworkInfo implementations by IAAS and CAAS.
This is skeletal, simply introducing the indirection, factory and
specific types. Separation of functionality as appropriate will follow.
