+++ apiserver/facades/agent/uniter/export_test.go
@@ -46,6 +46,6 @@ func PatchGetStorageStateError(patcher patcher, err error) {
 	patcher.PatchValue(&getStorageState, func(st *state.State) (storageAccess, error) { return nil, err })
 }
 
-func (n *NetworkInfo) MachineNetworkInfos(spaceIDs ...string) (map[string]machineNetworkInfoResult, error) {
+func (n *NetworkInfoBase) MachineNetworkInfos(spaceIDs ...string) (map[string]machineNetworkInfoResult, error) {
 	return n.machineNetworkInfos(spaceIDs...)
 }
