+++ apiserver/facades/agent/uniter/networkinfocaas.go
@@ -1,0 +1,9 @@
+// Copyright 2020 Canonical Ltd.
+// Licensed under the AGPLv3, see LICENCE file for details.
+
+package uniter
+
+// NetworkInfoCAAS is used to provide network info for CAAS units.
+type NetworkInfoCAAS struct {
+	*NetworkInfoBase
+}
